requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
psycopg2-binary>=2.9
django-cors-headers>=4.0
python-dotenv>=1.0
openpyxl>=3.1
requests>=2.31
